+++ lib/authrocket/named_permission.rb
@@ -1,0 +1,10 @@
+module AuthRocket
+  class NamedPermission < Resource
+    crud :find, :create, :update, :delete
+
+    belongs_to :realm
+
+    attr :auto_grant, :name, :permission
+
+  end
+end