+++ lib/authrocket/notification.rb
@@ -1,8 +1,8 @@
 module AuthRocket
   class Notification < Resource
 
-    belongs_to :app_hook
     belongs_to :event
+    belongs_to :hook
 
     attr :attempts, :hook_type, :last_destination, :last_result, :state
     attr_datetime :last_attempt_at