+++ lib/authrocket/oauth2_session.rb
@@ -1,0 +1,26 @@
+module AuthRocket
+  class Oauth2Session < Resource
+    crud :find, :create
+
+    attr :access_token, :code, :id_token, :redirect_uri
+    attr :profile
+    attr :expires_in, :token_type
+    attr_datetime :expires_at
+
+    def self.json_root ; 'session' ; end
+    def self.url ; 'sessions/oauth2' ; end
+
+    class << self
+
+      # params - {client_app_id:, client_app_secret:, code:}
+      # returns: Token - must check .valid? or .errors? for response
+      def code_to_token(params={})
+        params = parse_request_params(params, json_root: json_root)
+        parsed, creds = request(:post, "#{url}/code", params)
+        factory(parsed, creds)
+      end
+
+    end
+
+  end
+end