+++ lib/authrocket/token.rb
@@ -1,0 +1,9 @@
+module AuthRocket
+  class Token < Resource
+
+    belongs_to :user
+
+    attr :token
+
+  end
+end